// src/app/models/cake-request.ts
export type CakeRequest = {
  id?: number;
  cakeName?: string;
  customerName?: string;
  customerEmail?: string;
  customerPhone?: string;
  dateOfDelivery?: Date;
  street?: string;
  city?: string;
  state?: string;
  postalCode?: number;
  quantity?: number;
  totalPrice?: number;
  cakePrice?: number

  
};